feat: multiple changes
- use cyclopts for command line arguments
- shared rich library console instance
- use Cbh class instead of CbhNetCdf
- CBH source can now be netcdf or zarr
- cleanup messages
- cleanup handling of writing model output variables. Hardcoded segment variables removed.
