+++ Bandit/bandit.py
@@ -1,8 +1,7 @@
 #!/usr/bin/env python3
 
-import argparse
 import datetime
-import errno
+# import errno
 import logging
 import networkx as nx   # type: ignore
 import numpy as np
@@ -10,13 +9,13 @@ import os
 import sys
 import time
 
-from collections import OrderedDict
+from cyclopts import App, Parameter, validators
 from dask.distributed import Client
 from packaging.version import Version
-from typing import List
+from pathlib import Path
+from typing import Annotated, List, Optional, Union
 
-from rich.console import Console
-from rich import pretty
+from pyPRMS.base.console import get_console_instance
 
 from Bandit import __version__
 from Bandit.bandit_helpers import (parse_gages, set_date, subset_stream_network, get_hru_and_seg_subset_maps,
@@ -30,7 +29,7 @@ import Bandit.prms_nwis as prms_nwis   # type: ignore
 
 from pyPRMS.constants import HRU_DIMS, PRMS_VERSION   # type: ignore
 from pyPRMS.metadata.metadata import MetaData   # type: ignore
-from pyPRMS import CbhNetcdf   # type: ignore
+from pyPRMS import Cbh   # type: ignore
 from pyPRMS import ControlFile   # type: ignore
 from pyPRMS import ParamDb   # type: ignore
 from pyPRMS import Parameters   # type: ignore
@@ -46,8 +45,7 @@ warnings.filterwarnings('ignore', message=r'.*organizePolygons\(\) received a po
 # from pyogrio import list_drivers, list_layers, read_info, read_dataframe, write_dataframe
 
 # Rich library
-pretty.install()
-con = Console()
+con = get_console_instance(record=True)
 
 __author__ = 'Parker Norton (pnorton@usgs.gov)'
 
@@ -71,108 +69,85 @@ flog.setFormatter(log_fmt)
 root.addHandler(flog)
 bandit_log.addHandler(clog)
 
+app = App(default_parameter=Parameter(negative=()))
 
-def main():
-    # Command line arguments
-    parser = argparse.ArgumentParser(description='Extract model subsets from the National Hydrologic Model')
-    parser.add_argument('-O', '--output_dir', help='Output directory for subset')
-    parser.add_argument('-p', '--param_filename', help='Name of output parameter file')
-    parser.add_argument('-s', '--streamflow_filename', help='Name of streamflow data file')
-    parser.add_argument('-P', '--paramdb_dir', help='Location of parameter database')
-    parser.add_argument('-M', '--merged_paramdb_dir', help='Location of merged parameter database')
-    parser.add_argument('-C', '--cbh_dir', help='Location of CBH files')
-    parser.add_argument('-g', '--geodatabase_filename', help='Full path to NHM geodatabase')
-    parser.add_argument('-j', '--job', help='Job directory to work in')
-    parser.add_argument('-v', '--verbose', help='Output additional information', action='store_true')
-    parser.add_argument('--check_DAG', help='Verify the streamflow network', action='store_true')
-    parser.add_argument('--output_cbh', help='Output CBH files for subset', action='store_true')
-    parser.add_argument('--output_shapefiles', help='Output shapefiles for subset', action='store_true')
-    parser.add_argument('--output_streamflow', help='Output streamflows for subset', action='store_true')
-    parser.add_argument('--cbh_netcdf', help='Enable netCDF output for CBH files', action='store_true')
-    parser.add_argument('--param_netcdf', help='Enable netCDF output for parameter file', action='store_true')
-    parser.add_argument('--streamflow_netcdf', help='Enable netCDF output for streamflow file', action='store_true')
-    parser.add_argument('--add_gages', metavar="KEY=VALUE", nargs='+',
-                        help='Add arbitrary streamgages to POIs of form gage_id=segment. Segment must ' +
-                             'exist in the model subset. Additional streamgages are marked as poi_type=0.')
-    parser.add_argument('--no_filter_params',
-                        help='Output all parameters regardless of modules selected', action='store_true')
-    parser.add_argument('--keep_hru_order',
-                        help='Keep HRUs in the relative order they occur in the paramdb', action='store_true')
-    parser.add_argument('--hru_gis_id', help='Name of key/id for HRUs in geodatabase', nargs='?', type=str)
-    parser.add_argument('--seg_gis_id', help='Name of key/id for segments in geodatabase', nargs='?', type=str)
-    parser.add_argument('--hru_gis_layer', help='Name of geodatabase layer containing HRUs', nargs='?', type=str)
-    parser.add_argument('--seg_gis_layer', help='Name of geodatabase layer containing Segments', nargs='?', type=str)
-    parser.add_argument('--prms_version', help='Write PRMS version 5 or 6 parameter file', nargs='?',
-                        default=str(PRMS_VERSION), type=str)
-    args = parser.parse_args()
 
-    prms_version = Version(args.prms_version)
+@app.default
+def extract(config_file: Annotated[Path, Parameter(validator=validators.Path(exists=True))] = Path('bandit.cfg'),
+            job_dir: Union[str, Path] = None,
+            verbose: bool = False,
+            cbh_netcdf: bool = False,
+            model_output_netcdf: bool = False,
+            param_netcdf: bool = False,
+            streamflow_netcdf: bool = False,
+            no_filter_params: bool = False,
+            keep_hru_order: bool = False,
+            prms_version: Union[str, Version] = PRMS_VERSION,
+            add_gages: Optional[List[str]] = None):
+    """Extract model subsets from the National Hydrologic Model parameter database
+
+    :param config_file: Name of config file to use for extraction
+    :param job_dir: Name of job directory to work in
+    :param verbose: Output additional debugging information
+    :param cbh_netcdf: Output CBH forcings in netCDF format
+    :param model_output_netcdf: Output model output variables in netCDF format
+    :param param_netcdf: Output parameter file in netCDF format
+    :param streamflow_netcdf: Output streamflow observations in netCDF format
+    :param no_filter_params: Output all parameters regardless of the control file options
+    :param keep_hru_order: Keep the HRU order relative to the source parameter database order
+    :param prms_version: Output control variables and parameters supported by the version number of PRMS
+    :param add_gages: Include additional ad-hoc streamgages as POIs using the form gage_id=segment.
+                      Each gage_id, segment pair should be separated with a space. Segment must
+                      exist in the model subset and not be used by existing POIs. These ad-hoc gages
+                      are marked in the subset parameter file as poi_type=0
+    """
+
+    if isinstance(prms_version, str):
+        prms_version: Version = Version(prms_version)
 
     stdir = os.getcwd()
 
-    if args.job:
-        if os.path.exists(args.job):
+    if job_dir is not None:
+        if isinstance(job_dir, str):
+            job_dir = Path(job_dir)
+
+        if job_dir.is_dir():
             # Change into job directory before running extraction
-            os.chdir(args.job)
-            # print('Working in directory: {}'.format(args.job))
+            os.chdir(job_dir)
         else:
-            print(f'ERROR: Invalid jobs directory: {args.job}')
+            print(f'ERROR: Invalid jobs directory: {str(job_dir)}')
             exit(-1)
 
     bandit_log.info(f'========== START {datetime.datetime.now().isoformat()} ==========')
 
     addl_gages = None
-    if args.add_gages:
-        addl_gages = parse_gages(args.add_gages)
+    if add_gages:
+        addl_gages = parse_gages(add_gages)
+        if verbose:
+            con.print(f'Additional streamgages: {addl_gages}')
         bandit_log.info('Additionals streamgages specified on command line')
 
-    config = bc.Cfg('bandit.cfg')
+    config = bc.Cfg(config_file)
 
-    # Override configuration variables with any command line parameters
-    for kk, vv in args.__dict__.items():
-        if config.exists(kk):
-            if vv:
-                bandit_log.info(f'Overriding configuration for {kk} with {vv}')
-                config.update_value(kk, vv)
+    outdir = Path(config.output_dir)   # Where to output the subset
+    param_filename = Path(config.param_filename)   # Name of the output parameter file
+    paramdb_dir = Path(config.paramdb_dir)   # Location of the NHM parameter database
+    cbh_dir = Path(config.cbh_dir)
+    dsmost_seg = config.outlets   # List of outlets
+    uscutoff_seg = config.cutoffs   # List of upstream cutoffs
+    hru_noroute = config.hru_noroute   # List of additional HRUs (have no route to segment within subset)
 
-    # Where to output the subset
-    outdir = config.output_dir
-
-    # What to name the output parameter file
-    param_filename = config.param_filename
-
-    # Location of the NHM parameter database
-    paramdb_dir = config.paramdb_dir
-
-    # List of outlets
-    dsmost_seg = config.outlets
-
-    # List of upstream cutoffs
-    uscutoff_seg = config.cutoffs
-
-    # List of additional HRUs (have no route to segment within subset)
-    hru_noroute = config.hru_noroute
-
-    # if args.prms_version == 6:
-    #     args.cbh_netcdf = True
-    #     args.param_netcdf = True
-    #     args.streamflow_netcdf = True
+    # if prms_version == 6:
+    #     cbh_netcdf = True
+    #     param_netcdf = True
+    #     streamflow_netcdf = True
 
     # Load PRMS metadata
-    prms_meta = MetaData(version=prms_version, verbose=False).metadata
+    con.print(f'[green4]INFO[/]: Loading PRMS metadata for version {prms_version}')
+    prms_meta = MetaData(version=prms_version, verbose=verbose).metadata
 
     # Load the control file
-    ctl = ControlFile(config.control_filename, metadata=prms_meta)
-
-    if ctl.has_dynamic_parameters:
-        if config.dyn_params_dir:
-            if not os.path.exists(config.dyn_params_dir):
-                bandit_log.error(f'dyn_params_dir: {config.dyn_params_dir}, does not exist.')
-                exit(2)
-        else:
-            bandit_log.error('Control file has dynamic parameters but dyn_params_dir is not specified ' +
-                             'in the config file')
-            exit(2)
+    ctl = ControlFile(config.control_filename, metadata=prms_meta, verbose=verbose)
 
     # Date range for pulling NWIS streamgage observations and CBH data
     st_date = set_date(config.start_date)
@@ -186,26 +161,29 @@ def main():
     # Output revision of NhmParamDb
     git_url = git_commit_url(paramdb_dir)
     nhmparamdb_revision = git_commit(paramdb_dir, length=7)
-    bandit_log.info(f'Using parameter database from: {git_repo(paramdb_dir)}')
-    bandit_log.info(f'Repo branch: {git_branch(paramdb_dir)}')
-    bandit_log.info(f'Repo commit: {nhmparamdb_revision}')
+    bandit_log.info(f'Parameter database: {git_repo(paramdb_dir)}')
+    bandit_log.info(f'Branch: {git_branch(paramdb_dir)}')
+    bandit_log.info(f'Commit: {nhmparamdb_revision}')
 
     # client = Client(threads_per_worker=1)
     client = Client()
     dash_link = client.dashboard_link
-    print(f'Dask dashboard: {dash_link}')
+    con.print(f'Dask dashboard: {dash_link}')
 
     # Load the NHMparamdb
-    if args.verbose:
-        con.print('Loading NHM ParamDb', style='green4')
+    if verbose:
+        # con.print('[green4]INFO[/]: Loading NHM ParamDb')
+        con.print(f'[green4]INFO[/]: Parameter database: {git_repo(paramdb_dir)}')
+        con.print(f'[green4]INFO[/]: Branch: {git_branch(paramdb_dir)}')
+        con.print(f'[green4]INFO[/]: Commit: {nhmparamdb_revision}')
 
-    pdb = ParamDb(paramdb_dir=paramdb_dir, metadata=prms_meta, verbose=args.verbose)
+    pdb = ParamDb(paramdb_dir=paramdb_dir, metadata=prms_meta, verbose=verbose)
     pdb.control = ctl
 
     if pdb.dimensions.exists('npoigages') and not pdb.dimensions.exists('nobs'):
         # If we have poi gages and nobs is missing then add it
-        if args.verbose:
-            con.print('[gold3]Added missing nobs dimension[/]')
+        if verbose:
+            con.print('[gold3]WARNING[/]: Added missing nobs dimension')
         pdb.dimensions.add('nobs', size=pdb.dimensions.get('npoigages').size)
 
     # Add defaults for parameters that are missing but required for the selected modules
@@ -213,15 +191,18 @@ def main():
     #                         and the wrong number of HRUs will likely be output for the extraction.
     pdb.add_missing_parameters()
 
-    if not args.no_filter_params:
+    if not no_filter_params:
         # Reduce the parameters to those required by the selected modules
         pdb.remove(pdb.unneeded_parameters)
+
+    if not pdb.exists('poi_gage_segment'):
+        con.print('[gold3]WARNING[/]: Missing POI-related parameters. To include POIs, set csvON_OFF > 0 in the control file')
 
     # Default the various *ON_OFF variables to 0 (off)
     # The original values are needed to reduce parameters by module,
     # but it's best to disable them in the final control file since
     # no output variables are defined for them.
-    disable_vars = ['basinOutON_OFF', 'mapOutON_OFF', 'nhruOutON_OFF',
+    disable_vars = ['basinOutON_OFF', 'csvON_OFF', 'mapOutON_OFF', 'nhruOutON_OFF',
                     'nsegmentOutON_OFF', 'nsubOutON_OFF']
     for vv in disable_vars:
         ctl.get(vv).values = 0
@@ -233,8 +214,8 @@ def main():
     # Convert to list for fastest access to array
     nhm_seg = pdb.get('nhm_seg').tolist()
 
-    if args.verbose:
-        con.print('Generating stream network', style='green4')
+    # if verbose:
+    #     con.print('Generating stream network', style='green4')
 
     # First check if any of the requested stream segments exist in the NHM.
     # An intersection of 0 elements can occur when all stream segments are
@@ -243,6 +224,7 @@ def main():
     #       tosegment but does exist in nhm_seg (e.g. standalone segment). So
     #       we use nhm_seg to verify at least one of the given segment(s) exist.
     if dsmost_seg and len(set(dsmost_seg).intersection(nhm_seg)) == 0:
+        con.print(f'[red]ERROR[/]: None of the requested stream segments exist in the NHM')
         bandit_log.error('None of the requested stream segments exist in the NHM paramDb')
         exit(200)
 
@@ -254,22 +236,28 @@ def main():
 
     if config.check_DAG:
         if not nx.is_directed_acyclic_graph(dag_ds):
+            con.print('[red]ERROR[/]: Cycles and/or loops found in stream network')
             bandit_log.error('Cycles and/or loops found in stream network')
 
             for xx in nx.simple_cycles(dag_ds):
+                con.print(f'[red]ERROR[/]: Cycle found for segment {xx}')
                 bandit_log.error(f'Cycle found for segment {xx}')
 
-    if args.verbose:
-        con.print('    Extracting model subset', style='blue')
+    # if verbose:
+    #     con.print('    Extracting model subset', style='blue')
 
     dag_ds_subset = subset_stream_network(dag_ds, uscutoff_seg, dsmost_seg)
 
     # Segments in model subset
     new_nhm_seg = [ee[0] for ee in dag_ds_subset.edges]
-    bandit_log.info(f'Number of segments in subset: {len(new_nhm_seg)}')
+    con.print(f'[green4]INFO[/]: Number of stream segments in model subset: {len(new_nhm_seg)}')
+    bandit_log.info(f'Number of segments in model subset: {len(new_nhm_seg)}')
+    if verbose:
+        con.print(f'segments: {new_nhm_seg}')
 
     # Using a dictionary mapping nhm_seg to 1-based index for speed
-    new_nhm_seg_to_idx1 = OrderedDict((ss, ii+1) for ii, ss in enumerate(new_nhm_seg))
+    new_nhm_seg_to_idx1 = dict((ss, ii+1) for ii, ss in enumerate(new_nhm_seg))
+    # new_nhm_seg_to_idx1 = OrderedDict((ss, ii+1) for ii, ss in enumerate(new_nhm_seg))
 
     # Generate the renumbered local tosegments (1-based with zero being an outlet)
     new_tosegment = [new_nhm_seg_to_idx1[ee[1]] if ee[1] in new_nhm_seg_to_idx1
@@ -290,20 +278,22 @@ def main():
 
     if set(hru_to_seg.values()) == set(hru_noroute):
         # This occurs when there are no ROUTED HRUs for any of the stream segments
-        # An exit code of 200 prevents bandit_by_streamgage and bandit_multi_locations
-        # from ending all remaining extractions
+        con.print('[red]ERROR[/]: No HRUs associated with any of the segments')
         bandit_log.error('No HRUs associated with any of the segments; exiting.')
-        exit(200)
+        exit(2)
 
     # HRU-related parameters can either be output with the legacy, segment-oriented order
     # or can be output maintaining their original HRU-relative order from the parameter database.
     hru_order_subset, new_hru_segment = get_output_order(hru_to_seg, seg_to_hru, hru_segment,
                                                          nhm_id_to_idx, new_nhm_seg,
                                                          new_nhm_seg_to_idx1, hru_noroute,
-                                                         keep_hru_order=args.keep_hru_order)
+                                                         keep_hru_order=keep_hru_order)
 
+    con.print(f'[green4]INFO[/]: Number of HRUs in model subset: {len(hru_order_subset)}')
     bandit_log.info(f'Number of HRUs in subset: {len(hru_order_subset)}')
     bandit_log.info(f'Size of hru_segment for subset: {len(new_hru_segment)}')
+    if verbose:
+        con.print(f'HRUs: {hru_order_subset}')
 
     # Use hru_order_subset to pull selected indices for parameters with nhru dimensions
     # hru_order_subset contains the in-order indices for the subset of hru_segments
@@ -327,6 +317,8 @@ def main():
                                                                          seg_to_hru,
                                                                          addl_gages=addl_gages)
 
+    con.print(f'[green4]INFO[/]: Number of POI gages in model subset: {len(new_poi_gage_id)}')
+
     # ==================================================================
     # ==================================================================
     # Process the parameters and create a parameter file for the subset
@@ -334,6 +326,7 @@ def main():
 
     # Remove the POI-related parameters if we have no POIs
     if len(new_poi_gage_segment) == 0:
+        con.print('[gold3]WARNING[/]: No POIs found for model subset')
         bandit_log.warning('No POI gages found for subset; removing POI-related parameters.')
 
         for rp in ['poi_gage_id', 'poi_gage_segment', 'poi_type']:
@@ -418,22 +411,21 @@ def main():
         cnew_param.data = outdata
 
     # Write the new parameter file
-    if args.verbose:
+    if verbose:
         con.print(f'[green4]INFO[/]: Writing parameter file for PRMS {prms_version}')
 
     header = [f'Written by Bandit version {__version__} for PRMS {prms_version}',
               f'ParamDb revision: {git_url}']
-    if args.param_netcdf:
+    if param_netcdf:
         # TODO: 2023-11-13 PAN - add version info and prms version as global attributes
-        base_filename = os.path.splitext(param_filename)[0]
-        param_filename = f'{base_filename}.nc'
-        new_ps.write_parameter_netcdf(f'{outdir}/{param_filename}')
+        param_filename = Path(f'{param_filename.stem}.nc')
+        new_ps.write_parameter_netcdf(outdir / param_filename)
     else:
-        new_ps.write_parameter_file(f'{outdir}/{param_filename}', header=header)
+        new_ps.write_parameter_file(outdir / param_filename, header=header)
 
-    ctl.get('param_file').values = param_filename
+    ctl.get('param_file').values = str(param_filename)
 
-    if args.verbose:
+    if verbose:
         sys.stdout.write('\n')
     sys.stdout.flush()
 
@@ -445,122 +437,119 @@ def main():
         # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
         # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
         # Subset the cbh files for the selected HRUs
-        if args.verbose:
+        if verbose:
             con.print('Processing CBH files', style='green4')
 
         # Read the CBH source file
-        if os.path.splitext(config.cbh_dir)[1] == '.nc':
-            cbh_hdl = CbhNetcdf(src_path=config.cbh_dir, st_date=st_date, en_date=en_date,
-                                nhm_hrus=hru_order_subset)
-        elif os.path.splitext(config.cbh_dir)[1] == '.json':
-            cbh_hdl = CbhNetcdf(src_path=config.cbh_dir, st_date=st_date, en_date=en_date,
-                                nhm_hrus=hru_order_subset)
-        elif os.path.splitext(config.cbh_dir)[1] == '.zarr':
-            cbh_hdl = CbhNetcdf(src_path=config.cbh_dir, st_date=st_date, en_date=en_date,
-                                nhm_hrus=hru_order_subset)
+        if cbh_dir.suffix == '.nc':
+            cbh_hdl = Cbh(cbh_dir, metadata=prms_meta, engine='netcdf')
+        elif cbh_dir.suffix == '.zarr':
+            cbh_hdl = Cbh(cbh_dir, metadata=prms_meta, engine='zarr')
         else:
-            raise ValueError('Missing netcdf CBH files')
+            raise ValueError('Missing CBH files')
 
-        if args.cbh_netcdf:
-            cbh_outfile = f'{outdir}/cbh.nc'
+        # Add the global NHM IDs
+        cbh_hdl.set_nhm_id(pdb.get('nhm_id').data)
+
+        if cbh_netcdf:
+            cbh_outfile = outdir / 'cbh.nc'
 
             global_attrs = dict(bandit_version=__version__, paramdb_url=git_url)
             cbh_hdl.write_netcdf(cbh_outfile, variables=list(config.cbh_var_map.keys()),
-                                 global_attrs=global_attrs)
+                                 global_attrs=global_attrs, time_slice=slice(st_date, en_date),
+                                 hru_ids=hru_order_subset)
 
             # Set the control file variables for the CBH files
             for cfv in config.cbh_var_map.values():
-                ctl.get(cfv).values = os.path.basename(cbh_outfile)
+                ctl.get(cfv).values = cbh_outfile.name
 
         else:
             for cvar, cfv in config.cbh_var_map.items():
-                if args.verbose:
+                if verbose:
                     con.print(f'--- {cvar}')
 
                 cfile = ctl.get(cfv).values
-                cbh_hdl.write_ascii(cfile, variable=cvar)
+                cbh_hdl.write_ascii(cfile, variable=cvar, time_slice=slice(st_date, en_date), hru_ids=hru_order_subset)
 
     # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     # Write output variables
     # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-    # 2019-08-07 PAN: first prototype for extractions of output variables
     if config.include_model_output:
-        # TODO: 2020-03-12 PAN - this is brittle, fix it.
-        # TODO: 2022-06-30 PAN - add option to write netCDF format
-        seg_vars = ['seginc_gwflow', 'seginc_potet', 'seginc_sroff', 'seginc_ssflow',
-                    'seginc_swrad', 'segment_delta_flow', 'seg_gwflow', 'seg_inflow',
-                    'seg_lateral_inflow', 'seg_outflow', 'seg_sroff', 'seg_ssflow',
-                    'seg_upstream_inflow', 'seg_cum_gwflow', 'seg_cum_potet',
-                    'seg_cum_sroff', 'seg_cum_ssflow']
+        src_model_output = Path(config.output_vars_dir)
+        model_output_dir = outdir / 'model_output'
+        model_output_dir.mkdir(exist_ok=True)
 
-        try:
-            os.makedirs(f'{outdir}/model_output')
-            bandit_log.info('Creating directory model_output, for model output variables')
-        except OSError:
-            bandit_log.info('Using existing model_output directory for output variables')
+        model_output_ds = ModelOutput(filename=src_model_output)
 
-        for vv in config.output_vars:
-            if args.verbose:
-                sys.stdout.write('\r                                                  ')
-                sys.stdout.write(f'\rProcessing output variable: {vv} ')
-                sys.stdout.flush()
-
-            filename = f'{config.output_vars_dir}/{vv}.nc'
-
-            try:
-                if vv in seg_vars:
-                    mod_out = ModelOutput(filename=filename, varname=vv, startdate=st_date, enddate=en_date,
-                                          nhm_segs=new_nhm_seg)
-                    mod_out.write_csv(f'{outdir}/model_output')
-                else:
-                    mod_out = ModelOutput(filename=filename, varname=vv, startdate=st_date, enddate=en_date,
-                                          nhm_hrus=hru_order_subset)
-                    mod_out.write_csv(f'{outdir}/model_output')
-            except FileNotFoundError:
-                bandit_log.warning(f'Model output variable, {vv}, does not exist; skipping.')
+        for cvar in config.output_vars:
+            if model_output_netcdf:
+                model_output_ds.write_netcdf(filename=model_output_dir / f'{cvar}.nc',
+                                             variables=cvar,
+                                             time_slice=slice(st_date, en_date),
+                                             hru_ids=hru_order_subset,
+                                             seg_ids=new_nhm_seg)
+            else:
+                model_output_ds.write_csv(pathname=model_output_dir,
+                                          variables=cvar,
+                                          time_slice=slice(st_date, en_date),
+                                          hru_ids=hru_order_subset,
+                                          seg_ids=new_nhm_seg)
 
     # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     # Write dynamic parameters
     # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     if ctl.has_dynamic_parameters:
-        # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-        # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-        # Add dynamic parameters
-        for cparam in ctl.dynamic_parameters:
-            param_name = f'dyn_{cparam}'
-            input_file = f'{config.dyn_params_dir}/{param_name}.nc'
-            output_file = f'{outdir}/{param_name}.param'
+        if config.dyn_params_dir is None:
+            bandit_log.error('Control file has dynamic parameters but dyn_params_dir is not specified ' +
+                             'in the config file')
+            exit(2)
+        else:
+            dyn_params_dir = Path(config.dyn_params_dir)
 
-            if not os.path.exists(input_file):
-                warn_txt = f'WARNING: CONUS dynamic parameter file: {input_file}, does not exist... skipping'
-                bandit_log.warning(warn_txt)
-            else:
-                if args.verbose:
-                    print(f'Writing dynamic parameter {cparam}')
+            if not dyn_params_dir.is_dir():
+                bandit_log.error(f'dyn_params_dir: {config.dyn_params_dir}, does not exist.')
+                exit(2)
 
-                mydyn = dyn_params.DynamicParameters(input_file, cparam, st_date, en_date, hru_order_subset)
+            # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+            # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+            # Add dynamic parameters
+            for cparam in ctl.dynamic_parameters:
+                param_name = f'dyn_{cparam}'
+                input_file = dyn_params_dir / f'{param_name}.nc'
+                output_file = outdir / f'{param_name}.param'
+                # input_file = f'{config.dyn_params_dir}/{param_name}.nc'
+                # output_file = f'{outdir}/{param_name}.param'
 
-                mydyn.read_netcdf()
-                out_order = [kk for kk in hru_order_subset]
+                if not input_file.is_file():
+                    warn_txt = f'WARNING: CONUS dynamic parameter file: {input_file}, does not exist... skipping'
+                    bandit_log.warning(warn_txt)
+                else:
+                    if verbose:
+                        print(f'Writing dynamic parameter {cparam}')
 
-                for cc in ['day', 'month', 'year']:
-                    out_order.insert(0, cc)
+                    mydyn = dyn_params.DynamicParameters(str(input_file), cparam, st_date, en_date, hru_order_subset)
 
-                header = ' '.join(map(str, out_order))
+                    mydyn.read_netcdf()
+                    out_order = [kk for kk in hru_order_subset]
 
-                # Output ASCII files
-                out_ascii = open(output_file, 'w')
-                out_ascii.write(f'{cparam}\n')
-                out_ascii.write(f'{header}\n')
-                out_ascii.write('####\n')
-                mydyn.data.to_csv(out_ascii, columns=out_order, na_rep='-999',
-                                  sep=' ', index=False, header=False, encoding=None, chunksize=50)
-                out_ascii.close()
+                    for cc in ['day', 'month', 'year']:
+                        out_order.insert(0, cc)
+
+                    header = ' '.join(map(str, out_order))   # type: ignore
+
+                    # Output ASCII files
+                    out_ascii = open(output_file, 'w')
+                    out_ascii.write(f'{cparam}\n')
+                    out_ascii.write(f'{header}\n')
+                    out_ascii.write('####\n')
+                    mydyn.data.to_csv(out_ascii, columns=out_order, na_rep='-999',
+                                      sep=' ', index=False, header=False, encoding=None, chunksize=50)
+                    out_ascii.close()
 
     # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     # Write control file
     # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-    ctl.write(f'{os.path.basename(config.get_value("control_filename"))}.bandit')
+    ctl.write(str(Path(config.control_filename).with_suffix('.bandit')))
 
     # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     # Write streamflow
@@ -570,30 +559,32 @@ def main():
         # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
         # Download the streamgage information from NWIS
         if len(new_poi_gage_id) > 0:
-            if args.verbose:
-                con.print(f'Retrieving streamgage observations for {len(new_poi_gage_id)} stations', style='green4')
+            if verbose:
+                con.print(f'[green4]INFO[/]: Retrieving streamgage observations for {len(new_poi_gage_id)} stations', style='green4')
 
             if config.exists('poi_dir') and config.poi_dir != '':
                 bandit_log.info('Retrieving POIs from local HYDAT and NWIS netcdf files')
                 streamflow = POI(src_path=config.poi_dir, st_date=st_date, en_date=en_date,
-                                 gage_ids=new_poi_gage_id, verbose=args.verbose)
+                                 gage_ids=new_poi_gage_id, verbose=verbose)
             else:
                 # Default to retrieving only NWIS stations from waterservices.usgs.gov
                 bandit_log.info('No poi_dir: retrieving only NWIS POIs from online NWIS service.')
                 streamflow = prms_nwis.NWIS(gage_ids=new_poi_gage_id, st_date=st_date, en_date=en_date,
-                                            verbose=args.verbose)
+                                            verbose=verbose)
                 streamflow.get_daily_streamgage_observations()
 
-            if args.streamflow_netcdf:
-                streamflow.write_netcdf(filename=f'{outdir}/{config.streamflow_filename}.nc')
+            if streamflow_netcdf:
+                streamflow.write_netcdf(filename=outdir / f'{config.streamflow_filename}.nc')
             else:
-                streamflow.write_ascii(filename=f'{outdir}/{config.streamflow_filename}')
+                streamflow.write_ascii(filename=outdir / f'{config.streamflow_filename}')
         else:
-            if args.verbose:
-                con.print('No POIs exist in model subset; writing dummy data', style='gold3')
-            streamflow = prms_nwis.NWIS(gage_ids=None, st_date=st_date, en_date=en_date, verbose=args.verbose)
+            # TODO: 2025-03-25 PAN - this should write a netcdf file if that option was selected
+            if verbose:
+                con.print('[green4]WARNING[/]: No POIs exist in model subset; writing dummy data', style='gold3')
+            streamflow = prms_nwis.NWIS(gage_ids=None, st_date=st_date, en_date=en_date, verbose=verbose)
             streamflow.get_daily_streamgage_observations()
-            streamflow.write_ascii(filename=f'{config.streamflow_filename}')
+            streamflow.write_ascii(filename=outdir / f'{config.streamflow_filename}')
+            # streamflow.write_ascii(filename=f'{config.streamflow_filename}')
             bandit_log.info(f'No POIs exist in model subset; dummy data written.')
 
     # *******************************************
@@ -601,31 +592,28 @@ def main():
     if config.output_shapefiles:
         stime = time.time()
 
-        if args.verbose:
+        src_gis = Path(config.gis['src_filename'])
+
+        if verbose:
             print('-'*40)
             con.print('Writing shapefiles for model subset', style='green4')
 
-        if len(config.gis) == 0 or not os.path.exists(config.gis['src_filename']):
+        if len(config.gis) == 0 or not src_gis.exists():
             bandit_log.error(f'Source GIS file'
                              f'does not exist. Shapefiles will not be created')
         else:
             # Create GIS subdirectory if it doesn't already exist
-            gis_dir = f'{outdir}/GIS'
-            try:
-                os.makedirs(gis_dir)
-            except OSError as exception:
-                if exception.errno != errno.EEXIST:
-                    raise
-                else:
-                    pass
+            gis_dir = outdir / 'GIS'
+            gis_dir.mkdir(exist_ok=True)
 
-            geo_outfile = f'{gis_dir}/model_layers.{config.gis["dst_extension"]}'
+            dst_gis_type = config.gis["dst_extension"]
+            geo_outfile = gis_dir / f'model_layers.{dst_gis_type}'
 
             for kk, vv in config.gis['layers'].items():
                 vv['include_fields'].extend([vv['key']])
 
                 if vv['type'] == 'nhru':
-                    geo_file = pyg.read_dataframe(config.gis['src_filename'], layer=vv['layer'],
+                    geo_file = pyg.read_dataframe(src_gis, layer=vv['layer'],
                                                   columns=vv['include_fields'], force_2d=True,
                                                   where=f'{vv["key"]} >= {min(hru_order_subset)} AND {vv["key"]} <= {max(hru_order_subset)}')
                     bb = geo_file[geo_file[vv['key']].isin(hru_order_subset)]
@@ -640,17 +628,17 @@ def main():
                     domain_layer = bb2.dissolve(aggfunc={'nhm_id': 'count'})
                     domain_layer.rename(columns={'nhm_id': 'num_hrus'}, inplace=True)
 
-                    if config.gis["dst_extension"] == 'gpkg':
+                    if dst_gis_type == 'gpkg':
                         bb.to_file(geo_outfile, layer=vv['type'], driver='GPKG')
                         domain_layer.to_file(geo_outfile, layer='domain', driver='GPKG')
                     else:
-                        geo_outfile = f'{gis_dir}/model_{vv["type"]}.{config.gis["dst_extension"]}'
+                        geo_outfile = gis_dir / f'model_{vv["type"]}.{dst_gis_type}'
                         bb.to_file(geo_outfile)
 
-                        domain_outfile = f'{gis_dir}/model_domain.{config.gis["dst_extension"]}'
+                        domain_outfile = gis_dir / f'model_domain.{dst_gis_type}'
                         domain_layer.to_file(domain_outfile)
                 elif vv['type'] == 'nsegment':
-                    geo_file = pyg.read_dataframe(config.gis['src_filename'], layer=vv['layer'],
+                    geo_file = pyg.read_dataframe(src_gis, layer=vv['layer'],
                                                   columns=vv['include_fields'], force_2d=True,
                                                   where=f'{vv["key"]} >= {min(new_nhm_seg)} AND {vv["key"]} <= {max(new_nhm_seg)}')
                     bb = geo_file[geo_file[vv['key']].isin(new_nhm_seg)]
@@ -658,36 +646,40 @@ def main():
                     local_ids = new_ps.get_dataframe('nhm_seg').reset_index()
                     bb = bb.merge(local_ids, on='nhm_seg')
 
-                    if config.gis["dst_extension"] == 'gpkg':
+                    if dst_gis_type == 'gpkg':
                         bb.to_file(geo_outfile, layer=vv['type'], driver='GPKG')
                     else:
-                        geo_outfile = f'{gis_dir}/model_{vv["type"]}.{config.gis["dst_extension"]}'
+                        geo_outfile = gis_dir / f'model_{vv["type"]}.{dst_gis_type}'
                         bb.to_file(geo_outfile)
                 elif vv['type'] == 'npoigages':
                     if len(new_poi_gage_id) > 0:
-                        geo_file = pyg.read_dataframe(config.gis['src_filename'], layer=vv['layer'],
+                        geo_file = pyg.read_dataframe(src_gis, layer=vv['layer'],
                                                       columns=vv['include_fields'], force_2d=True)
 
                         bb = geo_file[geo_file[vv['key']].isin(new_poi_gage_id)]
                         bb = bb.rename(columns={vv['key']: 'gage_id', vv['include_fields'][0]: 'nhm_seg'})
 
-                        if config.gis["dst_extension"] == 'gpkg':
+                        if dst_gis_type == 'gpkg':
                             bb.to_file(geo_outfile, layer=vv['type'], driver='GPKG')
                         else:
-                            geo_outfile = f'{gis_dir}/model_{vv["type"]}.{config.gis["dst_extension"]}'
+                            geo_outfile = gis_dir / f'model_{vv["type"]}.{dst_gis_type}'
                             bb.to_file(geo_outfile)
                     else:
                         bandit_log.info('No POIs in model subset so POI GIS layer not written.')
                 else:
                     bandit_log.warning(f'Layer, {kk}, has unknown type, {vv["type"]}; skipping.')
 
-        if args.verbose:
-            print(f'Geo write time: {time.time() - stime:0.3f} s', flush=True)
+        if verbose:
+            con.print(f'[green4]INFO[/]: Geo write time: {time.time() - stime:0.3f} s')
 
     bandit_log.info(f'========== END {datetime.datetime.now().isoformat()} ==========')
 
     os.chdir(stdir)
 
 
+def main():
+    app()
+
+
 if __name__ == '__main__':
-    main()
+    app()
